CoreClimate name change and base class
Changed the class name from Heidi to CoreClimate

Created a base class that the fuzion class can inherit from.
